[UPDATE] 목표 API 관련 수정

+++ src/main/java/com/slide_todo/slide_todoApp/dto/goal/IndividualGoalTodoDTO.java
@@ -16,7 +16,7 @@ public class IndividualGoalTodoDTO {
     private Long memberId;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Integer progress;
+    private Double progress;
     private List<IndividualTodoDto> todos;
 
     @Getter
